chore :: 포폴 리스트 페이지 루트로 옮김
